Create savedata.php
The PHP gets the GPS payload data and parses out the time, lat, and long into the "lunadata.txt" file. This becomes a running log of the GPS data.  

IMPORTANT!
The first time the PHP creates the "lunadata.txt" file, you must edit it or the webpage will break. Remove the "," (comma) that appears at the top of the page before the first "{" containing the GPS coordinates. Producing the comma is an artifact on how we are generating the subsequent {data}

The PHP also formats the GPS data and sends it to the locations.json file so the Leaflet script can display it on the webpage. It currently writes the time in UTC and needs a conversion script.